Update configurations and test script.
New versions of arm-optimized-routines contain exp10, and the tests
exercise it. However, on Android the routines are tested through libm,
which does does not have exp10, therefore the exp10 tests shouldn't be
compiled or run. For this end, they are disabled and the test running
script is updated to ignore the exp10 files.

The SIMD tests are also explicitly disabled, as WANT_VMATH is no
longer enough to avoid them.

Change-Id: I328998f8737a00c374220946cca94b7863f99276